Lem issue 96 (#97)
* liftAuth & liftUnauth
close #96

* Refactor.
Remove newEnv, runEnv
Drop 'Unauth' from newUnauthEnv, runUnauth
Ergvein.Wallet.Monad.Env -> Ergvein.Wallet.Monad.Auth
