+++ wallet/src/Ergvein/Wallet.hs
@@ -3,6 +3,5 @@ module Ergvein.Wallet(
   ) where
 
 import Ergvein.Wallet.Main as X
-import Ergvein.Wallet.Monad.Env as X
 import Ergvein.Wallet.Monad.Unauth as X
 import Ergvein.Wallet.Settings as X
